+++ src/haskell/list/index.ts
@@ -1,0 +1,2 @@
+export { permutations } from './permutations.ts'
+export { transpose } from './transpose.ts'