+++ src/kotlin/collections/index.ts
@@ -1,0 +1,1 @@
+export { groupBy } from './groupBy.ts'