+++ src/kotlin/index.ts
@@ -1,0 +1,2 @@
+export * as collections from './collections/index.ts'
+export * as text from './text/index.ts'