+++ src/kotlin/text/index.ts
@@ -1,0 +1,1 @@
+export { chunked } from './chunked.ts'