+++ src/swift/String/index.ts
@@ -1,0 +1,2 @@
+export { replacingOccurrences } from './replacingOccurrences.ts'
+export { split } from './split.ts'