+++ src/swift/index.ts
@@ -1,0 +1,1 @@
+export * as String from './String/index.ts'